Add bang command to leverage rg and sed
This is very experimental at this stage, but has had very promising results.

When applying very large changes, the original approach underperformed and even crashed vim.

Results tested for changing the word "user" to "usuario" in a project with 4539 matches for "user" was done in ~0.5s with this strategy.

Right now it leverages `rg`, but it looks like `grep` and `ag` also have the flag `--files-with-matches`, so technically this should be possible to make command agnostic with something like:

```vim
if executable('rg')
  let l:grep.command = 'rg'
endif
```